Disable drop area for dragging over descendants
If the user wishes to move an entry to its own children, i.e. putting it
and its children on the same level in the tree, then he will have to
manually move all the children as siblings of their parent (or
ancestor, more generally). One reason behind this decision was that
during a drag, when the dragged entry has not-collapsed children, the
drag picture (in Chrome at least) shows the children too, which may be
counter intuitive, since the user might expect, that the children
(descendants, more generally) will be also moved on drop, but that's
nonsense