Don't allow drop as child on collapsed entry
B/c if we allowed it, the dropped entry won't be visible immediately
(the user would have to open the drop target first), so it'd be counter
intuitive. If we automatically opened the drop target, then it'd be all
too complex etc. So it's been decided to simply prevent the drop as a
child, if the drop target entry is collapsed